internalize retryPolicy mechanism into Configuration
